+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Changelog file [CHANGELOG.md](CHANGELOG.md).
+- Tests for `git` command wrapper using [shellspec](https://github.com/shellspec/shellspec).
 
 ## 0.1.0 - 2024-04-10